Change StandardLayer to not use EJML internally
I did some tests with vs without EJML:
In a six-layer transformer model, I timed the amount of time it took to
get to 100 epochs. My laptop was plugged in and on best performance mode
for all trials. Over five trials without EJML, the average time was
12.7s (stdev = 0.36). Over five trials with EJML, average time was 15.1s
(stdev = 0.35). I did a sixth trial of without EJML after the EJML
trials to make sure that the difference wasn't caused by thermal
throttling. This time was consistent with the above times.
It is quite possible that there is a EJML implementation that is better
than the naive non-EJML implementation; I couldn't find it.